+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Make `as_slice()/from_slice()` const.
     * Make `ArrayLength: 'static`
     * Replace `From<&[T]>` with `TryFrom<&[T]>`
+    * Use `typenum`'s `const-generics` feature for `const N: usize`-based `From` implementations between `[T; N]` and `GenericArray<T, N>`
     * `alloc` crate feature
         * Added `box_arr!` macro with the same syntax as `arr!`, but returns a `Box<GenericArray<T, N>>`
         * Moving between heap and stack